Updated reading the user-token from config
The Client._logged_in variable is now set on startup if the user-token variable is set in the
config.json file.

Also, the EightTracksAPI._user_token variable is set upon calling __init__ to the user token
defined in the config file.

The headers are updated to have the 'X-User-Token' only if the _user_token variable is defined